+++ server/src/main.rs
@@ -1,12 +1,15 @@
 use std::path::PathBuf;
-
+use std::sync::Arc;
+use anyhow::anyhow;
 use clap::Parser;
 use opentelemetry::global::meter;
-use opentelemetry::metrics::Counter;
+use opentelemetry::metrics::ObservableCounter;
 use opentelemetry_sdk::metrics::SdkMeterProvider;
 use service::Service;
 use tracing::error;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, Layer};
+use state_store::IndexifyState;
+use crate::http_objects::IndexifyAPIError;
 
 mod config;
 mod executors;
@@ -27,21 +30,51 @@ struct Cli {
 
 #[derive(Clone)]
 struct MetricsData {
-    counter: Counter<u64>,
-}
-
-impl Default for MetricsData {
-    fn default() -> Self { Self::new() }
+    unallocated_tasks: ObservableCounter<u64>
 }
 
 impl MetricsData {
-    pub fn new() -> MetricsData {
-        let counter = meter("").u64_counter("counter1")
-            .with_description("Simple counter for testin")
+    pub fn new(indexify_state: Arc<IndexifyState>) -> MetricsData {
+        let unallocated_tasks = meter("unallocated_tasks_counter")
+            .u64_observable_counter("unallocated_tasks")
+            .with_description("Counter that displays unallocated tasks in the server")
+            .with_callback({
+                let indexify_state = indexify_state.clone();
+
+                move |observer| {
+                    let unallocated_tasks = indexify_state.reader().unallocated_tasks();
+
+                    match unallocated_tasks {
+                        Ok(tasks) => {
+                            for task in tasks {
+                                let compute_graph =
+                                    indexify_state.reader().get_compute_graph(&task.namespace, &task.compute_graph_name)
+                                        .map_err(|_| IndexifyAPIError::internal_error(anyhow!("Unable to read metrics")))
+                                        .unwrap();
+
+                                let compute_graph = match compute_graph{
+                                    None => { continue; }
+                                    Some(x) => { x }
+                                };
+
+                                let node = compute_graph.nodes.get(&task.compute_fn_name).unwrap();
+
+                                let image_version = node.image_version();
+                                let image_name = node.image_name().to_string();
+
+                                let version_kv = opentelemetry::KeyValue::new("image_version", image_version.to_string());
+                                let name_kv = opentelemetry::KeyValue::new("image_name", image_name);
+                                observer.observe(1, &[version_kv, name_kv])
+                            }
+                        },
+                        Err(_) => {},
+                    }
+                }
+            })
             .init();
 
         MetricsData {
-            counter
+            unallocated_tasks
         }
     }
 }
